a lot of changes.  - Got the crypto page working and able to add cryptocurrencies.  - Fixed some stuff with the regular stocks
